Fixing serialization problems
Specifically, it can't serialize floats and was trying to serialize/set
constants and statics

+++ Assets/Harmonic/2D/BasicWASD.cs
@@ -77,7 +77,7 @@ public class BasicWASD : MonoBehaviour {
 			TransitionKit.instance.transitionWithDelegate( pixelater );
 		}
 
-        ClockText.text = HarmonicUtils.ClockFormat(StarshipEnvironment.Instance.SecondsTilOxygenRunsOut);
+        ClockText.text = HarmonicUtils.ClockFormat((float)StarshipEnvironment.Instance.SecondsTilOxygenRunsOut);
 	}
 
     private void ToggleTerminal()
